requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-groq>=0.1.0
ollama>=0.2.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
